+++ test/index.js
@@ -3,3 +3,4 @@ require('./extend');
 require('./arrays');
 require('./toJSON');
 require('./clone');
+require('./sync');