+++ test/sync.js
@@ -9,6 +9,44 @@ var Post = Model({
     title: String
 }, { url_root: '/posts' });
 
+test('should POST when creating', function(done) {
+    function sync(opt, cb) {
+        assert.equal(opt.method, 'POST');
+        assert.equal(opt.url, '/posts');
+        assert.equal(opt.query, undefined);
+        cb(null, {
+            id: '12345'
+        });
+    };
+
+    var post = Post.extend({}, { sync: sync })();
+    post.save(function(err) {
+        assert.ifError(err);
+        assert.equal(post.id, 12345);
+        done();
+    });
+});
+
+test('should capture returned arguments from POST', function(done) {
+    function sync(opt, cb) {
+        assert.equal(opt.method, 'POST');
+        assert.equal(opt.url, '/posts');
+        assert.equal(opt.query, undefined);
+        cb(null, {
+            id: '12345',
+            title: 'hello world'
+        });
+    };
+
+    var post = Post.extend({}, { sync: sync })();
+    post.save(function(err) {
+        assert.ifError(err);
+        assert.equal(post.id, 12345);
+        assert.equal(post.title, 'hello world');
+        done();
+    });
+});
+
 test('destroy', function(done) {
     function sync(opt, cb) {
         assert.equal(opt.method, 'DELETE');